Created simple plugin scheme for GKCode
Much more work will be required in this area. Currently GKCode
classes are all treated as readers, and all have had 'verify'
functions added. However, they do much more than just read,
and a lot of their functionality should be moved elsewhere in
the project. For example, functions that barely touch 'self',
but do modify a Pyro object as a side effect should probably
be refactored to return something, and the Pyro object should
be tasked with updated its attributes.