Expanded unit tests for GKOutputReaderGS2
Currently parametrizing the mock_reader fixture, so all tests which use
it will be run on every combination of linearity/fields/fluxes. Further
tests should be added to account for when len(kx)=len(ky)=1, as then we
can also test if eigenvalues/eigenfunctions are being generated
correctly. The current parametrization is perhaps a bit wasteful, as
there's no need to retest _set_fluxes when there are no fields present,
no need to retest _set_fields when nonlinear, etc.