Ensure we don't write possibly unconverted values or NaN to file
If we convert a value and it becomes NaN, that indicates we don't have
a physical unit for the equivalent simulation unit. So we should
error, to avoid passing on values in the wrong units.

Currently special cases zero, because this usually indicates a value
is unset -- perhaps we should make sure we use None instead?