Add ConstNormalisation class, begin LocalGKSimulation
Adds a variant on both SimulationNormalisation and
ConventionNormalisation that provides very similar behaviour,
but without the option to redefine units after creation.
Also avoids splitting funcitonality between two classes that
each depend on each other.

Started work on a class to bundle together LocalGeometry,
LocalSpecies, Numerics, and a Normalisation. The aim is for
this to be an immutable representation of a single simulation,
and modifications to it will generate a new instance.