[TESTS FAIL] Rearrange from_gk_data
Fails when delta/s_delta are passed to create MXH, as it can't handle
the redundant values.

Planning to replace this with a non-classmethod utility function
which will be called by each subclass. Calls to from_gk_data will
be replaced with simple __init__ calls, and MXH.__init__ can
handle the edge case individually.