Replace from_gk_data with _init_with_shape_params
This is used by from_global_eq and from_local_geometry, as well
as all subclass __init__ functions. Provides a single unified
method for setting up a LocalGeometry after fitting shape
parameters.

Problem: Had to introduce variable overwrite_dpsidr, as previous
behaviour was to accept dpsidr when provided from an equilibrium
or a local geometry, but to always overwrite it when reading from
a GK input file. This feels like a bug, and the correct behaviour
should be to do this only if dpsidr is not provided, but it the
behaviour has been maintained in order to keep the tests correct.